Changed needed to allow Visual Studio 2010 to compile library.

+++ mythtv/libs/libmythtv/mpeg/premieredescriptors.cpp
@@ -48,8 +48,8 @@ QDateTime PremiereContentTransmissionDescriptor::StartTimeUTC(uint index) const
     // "Specification for Service Information in Digital Video Broadcasting"
     // to convert from Modified Julian Date to Year, Month, Day.
 
-    const float tmpA = 1.0 / 365.25;
-    const float tmpB = 1.0 / 30.6001;
+    const float tmpA = (float)(1.0 / 365.25);
+    const float tmpB = (float)(1.0 / 30.6001);
 
     float mjdf = mjd;
     int year  = (int) truncf((mjdf - 15078.2f) * tmpA);
